food catalog added

+++ lib/pages/food_pages/food_search_screen.dart
@@ -115,7 +115,7 @@ class _FoodSearchScreenState extends State<FoodSearchScreen> {
                             )
                           ]),
                       onTap: () {
-                        int sliderAmount = food.gramWeight;
+                        int sliderAmount = food.gramWeight.toInt();
                         showModalBottomSheet(
                             context: context,
                             builder: (context) => Center(
